Implemented a bot
Now the player is able to play on single or multi player. At first, the player must select the game mode (single or multi player). The single player mode is using a bot on 3 difficulties. The easy bot randomly chooses an available position, the medium bot tries to complete a line/diagonal/column to win the game or choose random and the hard bot firstly tries to win the game, next it tries to block the player to win and last choose random.